Removed apply and clear function
Apply function was unnecessary as it was essentially identical to how
the reset function worked. Therefore, apply function code was combined
with changeSquares event listener to dynamically change the number of
squares when using the slider.

Clear function was also unnecessary as it was always used in conjunction
with createNewSquares. Therefore, the code was combined with
createNewSquares function.